changed all components, split up Pagify component, changed index.js process.env
