+++ client/src/components/Builder/Builder.css
@@ -1,0 +1,25 @@
+.Builder {
+  width: 500px;
+  padding: 10px;
+  margin: 1rem;
+  box-shadow: 7px 10px 12px -5px;
+  text-align: center;
+  background-color: rgb(144, 144, 144);
+  border-radius: 3px;
+  float: left;
+  overflow: hidden;
+  margin-top: 5%;
+  margin-left: 10%;
+  font-family: 'Grenze', sans-serif;
+}
+
+.Builder-attributes {
+  color: whitesmoke;
+}
+
+.Builder-title {
+  margin-bottom: 1rem;
+  font-weight: bold;
+  color: #B71C1C;
+  font-size: 2.5rem;
+}
