+++ client/src/components/Buttons/Buttons.css
@@ -1,0 +1,3 @@
+button {
+    margin: .3em;
+}
