+++ client/src/components/Buttons/index.jsx
@@ -1,0 +1,1 @@
+export { default } from "./Buttons";
