+++ client/src/components/Character/Character.css
@@ -1,0 +1,28 @@
+.Character {
+    background-color: rgb(144, 144, 144);
+    border-radius: 3px;
+    width: 500px;
+    padding: 10px;
+    margin: 1rem;
+    box-shadow: 7px 10px 12px -5px;
+    text-align: center;
+    float: right;
+    overflow: hidden;
+    margin-top: 5%;
+    margin-left: 10;
+}
+
+.Character-image {
+    justify-content: center;
+    align-items: center;
+    border-radius: 3px;
+    margin-bottom: 1rem;
+}
+
+.Character-title {
+    margin-bottom: 1rem;
+    font-weight: bold;
+    font-size: 4rem;
+    color: #B71C1C;
+    font-family: 'Qwigley'
+}
