+++ client/src/components/NotFound/NotFound.jsx
@@ -1,4 +1,6 @@
 import React from 'react';
+import "./NotFound.css";
+
 
 function NotFound() {
     return <div className="container grey lighten-5">
