+++ client/src/components/Pagify/Page.jsx
@@ -1,0 +1,11 @@
+import React from 'react';
+
+const Page = () => {
+    return (
+        <React.Fragment>
+            Pager
+        </React.Fragment>
+    );
+};
+
+export default Page;
