+++ client/src/components/Pagify/index.jsx
@@ -1,0 +1,2 @@
+export { default as Pager } from "./Pager";
+export { default as Page } from "./Page";
