+++ client/src/components/Sign-up/index.jsx
@@ -1,0 +1,1 @@
+export { default } from "./Sign-up";
