+++ client/src/pages/CampaignCreator.jsx
@@ -1,6 +1,5 @@
 import React, {Component} from 'react';
-import {Pager, Page} from '../components/Pagify';
-
+import { Pager, Page } from "../components/Pagify";
 class CreateCampaign extends Component {
     render(){
         return (
