+++ package.json
@@ -23,6 +23,8 @@
     "express": "^4.16.3",
     "if-env": "^1.0.4",
     "mongoose": "^5.7.1",
-    "passport": "^0.4.0"
+    "passport": "^0.4.0",
+    "react": "^16.10.2",
+    "react-redux": "^7.1.1"
   }
 }
