+++ routes/index.js
@@ -1,12 +1,13 @@
-const path = require('path');
-const router = require('express').Router();
-const apiRoutes = require('./api');
+const path = require("path");
+const router = require("express").Router();
+const apiRoutes = require("./api");
 
+router.use("/api", apiRoutes);
 
-router.use('/api', apiRoutes);
-
-router.use((req, res) => {
-  res.sendFile(path.join(__dirname, '../client/build/index.html'))
-});
+if (process.env.NODE_ENV === "production") {
+  router.use((req, res) => {
+    res.sendFile(path.join(__dirname, "../client/build/index.html"));
+  });
+}
 
 module.exports = router;
