+++ server.js
@@ -23,9 +23,11 @@ app.use(router)
 mongoose.connect(process.env.MONGODB_URI || "mongodb://localhost/DungeonsAndDragonsBeotches");
 // Send every other request to the React app
 // Define any API routes before this runs
-app.get("*", (req, res) => {
-  res.sendFile(path.join(__dirname, "./client/public/index.html"));
-});
+if (process.env.NODE_ENV === 'production') {
+  app.get("*", (req, res) => {
+    res.sendFile(path.join(__dirname, "./client/public/index.html"));
+  });
+}
 
 app.listen(PORT, () => {
   console.log(`🌎 ==> API server now on port ${PORT}!`);
